src/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.0.9
django-allauth==65.3.0
django-allauth-ui==1.5.3
django-environ==0.11.2
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
Markdown==3.7
oauthlib==3.2.2
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
slippers==0.6.2
sqlparse==0.5.2
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
